The new UI layout and design implementation
* new UI
* Pug deprecated - no pages use it anymore
* MVC controllers architecture introduced
* Handelbars introduction
