+++ package.json
@@ -33,9 +33,13 @@
     "body-parser": "~1.18.2",
     "cookie-parser": "~1.4.3",
     "debug": "~2.6.9",
+    "ejs": "^2.5.7",
     "eslint": "^4.16.0",
+    "exphbs": "^1.1.0",
     "express": "~4.15.5",
+    "express-handlebars": "^3.0.0",
     "form-data": "^2.3.1",
+    "html": "^1.0.0",
     "insecure": "0.0.4",
     "mongodb": "^3.0.1",
     "mongoose": "^4.13.9",
@@ -51,7 +55,6 @@
     "xmljson": "^0.2.0"
   },
   "devDependencies": {
-    "babel": "^6.23.0",
     "babel-cli": "^6.26.0",
     "babel-istanbul": "^0.12.2",
     "babel-plugin-transform-runtime": "^6.23.0",
