+++ src/routes/index.js
@@ -3,7 +3,7 @@ const router = express.Router();
 
 /* GET home page. */
 router.get('/', function(req, res, next) {
-  res.render('index', { title: 'Home page' });
+    res.redirect("/web/index");
 });
 
 module.exports = router;
