+++ src/views_html/error.html
@@ -1,0 +1,1 @@
+This is error
