+++ src/views_html/includes/menu.html
@@ -1,0 +1,26 @@
+<!-- HEADER -->
+<header class="block">
+    <ul class="header-menu horizontal-list">
+        <li>
+            <a class="header-menu-tab" href="/web/letters"><span
+                    class="icon fontawesome-envelope scnd-font-color"></span>Messages</a>
+            <a class="header-menu-number" href="#4">24</a>
+        </li>
+        <li>
+            <a class="header-menu-tab" href="#2"><span class="icon fontawesome-user scnd-font-color"></span>Account</a>
+        </li>
+        <li>
+            <a class="header-menu-tab" href="/web/statistics"><span class="icon entypo-chart-line scnd-font-color"></span>Statistics</a>
+        </li>
+        <li>
+            <a class="header-menu-tab" href="#1"><span class="icon entypo-cog scnd-font-color"></span>Settings</a>
+        </li>
+    </ul>
+    <div class="profile-menu">
+        <p>Me <a href="#26"><span class="entypo-down-open scnd-font-color"></span></a></p>
+        <div class="profile-picture small-profile-picture">
+            <img width="40px" alt="Anne Hathaway picture"
+                 src="http://upload.wikimedia.org/wikipedia/commons/e/e1/Anne_Hathaway_Face.jpg">
+        </div>
+    </div>
+</header>
