+++ src/views_html/includes/statistics-menu.html
@@ -1,0 +1,34 @@
+<div class="left-container container">
+    <div class="menu-box block"> <!-- MENU BOX (LEFT-CONTAINER) -->
+        <h2 class="titular">Totally different menu</h2>
+        <ul class="menu-box-menu">
+            <li>
+                <a class="menu-box-tab" href="#6"><span class="icon fontawesome-envelope scnd-font-color"></span>Messages
+                    to verify
+                    <div class="menu-box-number">24</div>
+                </a>
+            </li>
+            <li>
+                <a class="menu-box-tab" href="/web/ticket"><span class="icon entypo-calendar scnd-font-color"></span>Get
+                    Ticket</a>
+            </li>
+            <li>
+                <a class="menu-box-tab" href="/web/letters"><span
+                        class="icon entypo-paper-plane scnd-font-color"></span>Sync with post</a>
+            </li>
+            <li>
+                <a class="menu-box-tab" href="/web/send-letter"><span
+                        class="icon entypo-paper-plane scnd-font-color"></span>Send letter
+                </a>
+            </li>
+            <li>
+                <a class="menu-box-tab" href="#12"><span class="icon entypo-cog scnd-font-color"></span>Letters Settings</a>
+            </li>
+            <li>
+                <a class="menu-box-tab" href="#13">
+                    <sapn class="icon entypo-chart-line scnd-font-color"></sapn>
+                    Statistics</a>
+            </li>
+        </ul>
+    </div>
+</div>
