+++ src/views_html/index.html
@@ -1,0 +1,9 @@
+<div class="join-newsletter block"> <!-- JOIN NEWSLETTER (RIGHT-CONTAINER) -->
+    <h2 class="titular">Index page</h2>
+    <div class="input-container">
+        <p>
+            This is the index page and who know how long this page can be? Can we write the text into the impossible
+            length?
+        </p>
+    </div>
+</div>
