+++ src/www.js
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 require("babel-register");
-require("babel-polyfill");
+//require("babel-polyfill");
 /**
  * Module dependencies.
  */
